requirements.txt
sentence-transformers
faiss-cpu
PyPDF2
pandas
requests
tqdm
matplotlib
scikit-learn
umap-learn
geopy
folium
numpy